requirements.txt
# Telegram Bot
python-telegram-bot>=20.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
